Extracts InOutOctree's MeshWrapper class into a separate file
